處理音檔轉換文字功能實現

+++ src/main/java/xyz/dowob/audiototext/entity/TranscriptionSegment.java
@@ -13,6 +13,6 @@ import lombok.Data;
 @Data
 public class TranscriptionSegment {
     private String text;
-    private Long startTime;
-    private Long endTime;
+    private Double startTime;
+    private Double endTime;
 }
